@@ -1,6 +1,6 @@
 using Odin.DesignContracts;
 
-namespace TargetsUntooled
+namespace Targets
 {
     /// <summary>
     /// The rewriter is expected to inject
@@ -25,10 +25,23 @@ namespace TargetsUntooled
         {
             _value++;
         }
+        
+        public async Task<int> AsyncIncrement()
+        {
+            _value++;
+            return await Task.FromResult(_value);
+        }
+
 
         public void MakeInvalid()
         {
             _value = -1;
+        }
+        
+        public async Task<int> AsyncMakeInvalid()
+        {
+            _value = -1;
+            return await Task.FromResult(_value);
         }
 
         [global::Odin.DesignContracts.Pure]
